+++ libs/core/src/cms/utils/getAccountNotifications.ts
@@ -5,7 +5,7 @@ import ms from 'ms.macro'
 import { NotificationModel } from '../types'
 import { getCmsClient } from '../utils'
 
-const NOTIFICATIONS_TTL = ms`5m` - ms`10s` // Substract 10s to not hit cache with polling in useAccountNotifications
+const NOTIFICATIONS_TTL = ms`5m` - ms`10s` // Subtract 10s to not hit cache with polling in useAccountNotifications
 
 const cache = new TTLCache<NotificationModel[]>('cmsAccountNotifications:v0', true, NOTIFICATIONS_TTL)