+++ libs/events/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [3.1.0](https://github.com/cowprotocol/cowswap/compare/events-v3.0.1...events-v3.1.0) (2026-02-05)
+
+
+### ✨ Features
+
+* **analytics:** 1 - enhance order lifecycle analytics ([#6402](https://github.com/cowprotocol/cowswap/issues/6402)) ([4e12fbf](https://github.com/cowprotocol/cowswap/commit/4e12fbf3a78a777445fbf5841c79ce5f14afb815))
+* **trade:** upload appData to IPFS before signing ([#6922](https://github.com/cowprotocol/cowswap/issues/6922)) ([cb94104](https://github.com/cowprotocol/cowswap/commit/cb9410470c549bd5453a7933a0a224ba706ec2c2))
+
+
+### 🔧 Miscellaneous
+
+* update sdk ([a20814b](https://github.com/cowprotocol/cowswap/commit/a20814b7caf16f495cd0292d5f3469bc552e4d97))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/types bumped to 3.1.0
+
 ## [3.0.1](https://github.com/cowprotocol/cowswap/compare/events-v3.0.0...events-v3.0.1) (2026-02-02)